--- __tests__/functional/core-marketplace/helper.ts
@@ -1,17 +1,0 @@
-import { TransactionFactory } from "../../../../__tests__/helpers";
-import {
-    BridgechainRegistrationBuilder,
-    BridgechainResignationBuilder,
-    BridgechainUpdateBuilder,
-    BusinessRegistrationBuilder,
-    BusinessResignationBuilder,
-    BusinessUpdateBuilder,
-} from "../../src/builders";
-import {
-    IBridgechainRegistrationAsset,
-    IBridgechainUpdateAsset,
-    IBusinessRegistrationAsset,
-    IBusinessUpdateAsset,
-} from "../../src/interfaces";
-
-export class MarketplaceTransactionFactory extends TransactionFactory {}
