+++ __tests__/unit/core-marketplace/builders/bridgechain-registration.test.ts
@@ -2,7 +2,7 @@ import "jest-extended";
 
 import { Managers, Transactions, Utils } from "@arkecosystem/crypto";
 import { BridgechainRegistrationBuilder } from "../../../../packages/core-marketplace/src/builders";
-import { MarketplaceTransactionType } from "../../../../packages/core-marketplace/src/marketplace-transactions";
+import { MarketplaceTransactionType } from "../../../../packages/core-marketplace/src/enums";
 import { BridgechainRegistrationTransaction } from "../../../../packages/core-marketplace/src/transactions";
 import { bridgechainRegistrationAsset1, bridgechainRegistrationAsset2 } from "../helper";
