+++ __tests__/unit/core-marketplace/builders/business-registration.test.ts
@@ -2,7 +2,7 @@ import "jest-extended";
 
 import { Managers, Transactions, Utils } from "@arkecosystem/crypto";
 import { BusinessRegistrationBuilder } from "../../../../packages/core-marketplace/src/builders";
-import { MarketplaceTransactionType } from "../../../../packages/core-marketplace/src/marketplace-transactions";
+import { MarketplaceTransactionType } from "../../../../packages/core-marketplace/src/enums";
 import { BusinessRegistrationTransaction } from "../../../../packages/core-marketplace/src/transactions";
 
 let builder: BusinessRegistrationBuilder;
